fix: [#22] Resolve circular dependencies
