+++ src/ValueProcessors/Generators/index.ts
@@ -1,7 +1,5 @@
 export * from "./Timestamp";
 export * from "./GetWindowSize";
 export * from "./GetMousePosition";
-export * from "./Const";
-export * from "./Var";
 export * from "./Random";
 export * from "./GetScale";
