+++ webpack.config.js
@@ -1,4 +1,5 @@
 const CleanWebpackPlugin = require('clean-webpack-plugin');
+const CircularDependencyPlugin = require('circular-dependency-plugin')
 const path = require("path");
 
 module.exports = {
@@ -32,6 +33,17 @@ module.exports = {
 	},
 	plugins: [
 		new DtsBundlePlugin(),
+		new CircularDependencyPlugin({
+			// exclude detection of files based on a RegExp
+			exclude: /a\.js|node_modules/,
+			// add errors to webpack instead of warnings
+			failOnError: true,
+			// allow import cycles that include an asyncronous import,
+			// e.g. via import(/* webpackMode: "weak" */ './file.js')
+			allowAsyncCycles: false,
+			// set the current working directory for displaying module paths
+			cwd: process.cwd(),
+		}),
 		new CleanWebpackPlugin({
 			verbose: true,
 			cleanOnceBeforeBuildPatterns: [],
